Refactor form handling and component structure across various routes
- Updated issue-related routes to use new component paths for modals.
- Added TypeScript support to new-house creation and joining forms.
- Simplified server actions by removing redundant user
authentication checks.
- Introduced new schemas for chat, house, and join code forms.
- Reorganized onboarding form structure and validation.
- Enhanced chat page with real-time updates and improved message rendering.
- Created reusable components for house and join code forms.
- Improved error handling and page structure for better user experience.
