+++ src/lib/form_schemas/chat/create_chat.schema.ts
@@ -1,0 +1,5 @@
+import { z } from 'zod';
+
+export const schema = z.object({
+	chat: z.string()
+});
