+++ src/lib/ui/components/Hero3DText.svelte
