--- src/lib/ui/pages/Houses.svelte
@@ -1,1 +1,0 @@
-<p>This is a test</p>
