+++ src/routes/(app)/new-house/join/+page.svelte
@@ -1,31 +1,12 @@
 <script lang="ts">
+	import UsejoinCodeForm from '$ui/components/join_code/UseJoinCode.form.svelte';
 	import type { PageProps } from './$types';
-	import { superForm } from 'sveltekit-superforms';
-	import { Field, Control, Label, FieldErrors } from 'formsnap';
-	import { zodClient } from 'sveltekit-superforms/adapters';
-	import { schema } from '$lib/form_schemas/join_code.schema.js';
 	import { page } from '$app/state';
-	let { data }: PageProps = $props();
-	const code = $derived(page.url.searchParams.get('code'));
 
-	const form = superForm(data.form, {
-		validators: zodClient(schema)
-	});
-	const { form: formData, enhance } = form;
-	$effect(() => {
-		$formData.join_code = code ?? '';
+	let { data }: PageProps = $props();
+	const { code, form } = $derived.by(() => {
+		return { code: page.url.searchParams.get('code'), form: data.form };
 	});
 </script>
 
-<form use:enhance class="mx-auto flex max-w-md flex-col" method="POST">
-	<Field {form} name="join_code">
-		<Control>
-			{#snippet children({ props })}
-				<Label>Join Code</Label>
-				<input {...props} type="text" bind:value={$formData.join_code} />
-			{/snippet}
-		</Control>
-		<FieldErrors />
-	</Field>
-	<button>Submit</button>
-</form>
+<UsejoinCodeForm {form} {code} />
