+++ src/routes/+error.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import Error from '$ui/pages/Error.svelte';
+	import Error from '$ui/pages/Error.page.svelte';
 	import { page } from '$app/state';
 	let { error, status }: { error: App.Error | null; status: number } = $derived(page);
 </script>
